Add memset_l(), memset32() and memset64()
memset_l() is like memset() but allows the user to fill the destination
with a pattern which fits in an unsigned long.  memset32() and memset64()
are 32-bit and 64-bit variants of this; memset_l() will call the
appropriate one.  memset32() is also useful by itself, while I am not
currently aware of any user who would benefit from memset64() on 32-bit.

Architecture maintainers should be able to write optimised versions of
these quite easily as it would share a lot with the core of an optimised
memset().

Signed-off-by: Matthew Wilcox <mawilcox@microsoft.com>

Signed-off-by: Paul Reioux <reioux@gmail.com>