arm64: csum: export symbol ip_fast_csum for modules
hm.. there are some out-of-tree kernel modules from broadcom not
100% compliant to GPL and needed ip_fast_csum for one of their
"pre-built" kernel modules

Instead of nice inline, this is separated out as a function call.
Not sure about the performance hit, but call vs inlining probably not as
efficient as it could be.

hopefully it's net positive compared to using standard non-aarch64
optimized library provided by the kernel

Signed-off-by: Paul Reioux <reioux@gmail.com>